add uri template support
